Move groupResults and resultMatchesCellFilter
