+++ packages/scrabble-solver/src/lib/index.ts
@@ -11,7 +11,6 @@ export { findCell } from './findCell';
 export { getCoordinate } from './getCoordinate';
 export { getCoordinates } from './getCoordinates';
 export { getTileSizes } from './getTileSizes';
-export { groupResults } from './groupResults';
 export { inverseDirection } from './inverseDirection';
 export { isCtrl } from './isCtrl';
 export { isMac } from './isMac';
@@ -21,7 +20,6 @@ export { localeTransliterate } from './localeTransliterate';
 export { memoize } from './memoize';
 export { noop } from './noop';
 export { numberComparator } from './numberComparator';
-export { resultMatchesCellFilter } from './resultMatchesCellFilter';
 export { reverseComparator } from './reverseComparator';
 export { sortGroupedResults } from './sortGroupedResults';
 export { sortResults } from './sortResults';
